move search and library
