--- app/src/main/java/com/example/playlistmaker/data/dto/TrackListResponse.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.data.dto
-
-data class TrackListResponse(
-    val resultCount: Int,
-    val results: MutableList<TrackDto>?
-) : Response()
-
