--- app/src/main/java/com/example/playlistmaker/domain/api/TrackRepository.kt
@@ -1,8 +1,0 @@
-package com.example.playlistmaker.domain.api
-
-import com.example.playlistmaker.domain.models.Track
-
-interface TrackRepository {
-
-    fun getTrack(): Track
-}
