--- app/src/main/java/com/example/playlistmaker/domain/api/TracksRepository.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.domain.api
-
-import com.example.playlistmaker.domain.models.Track
-
-interface TracksRepository {
-    fun searchTracks(expression: String): TracksData<List<Track>>
-}
