--- app/src/main/java/com/example/playlistmaker/domain/use_case/TracksInteractor.kt
@@ -1,13 +1,0 @@
-package com.example.playlistmaker.domain.use_case
-
-import com.example.playlistmaker.domain.api.TracksData
-import com.example.playlistmaker.domain.models.Track
-
-interface TracksInteractor {
-
-    fun searchTracks(expression: String, consumer: TrackConsumer)
-
-    interface TrackConsumer {
-        fun consume(data: TracksData<List<Track>>)
-    }
-}
