+++ app/src/main/java/com/example/playlistmaker/new/search/data/dto/TrackListResponse.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.new.search.data.dto
+
+import com.example.playlistmaker.data.dto.TrackDto
+
+data class TrackListResponse(
+    val resultCount: Int,
+    val results: MutableList<TrackDto>?
+) : Response()
