+++ app/src/main/java/com/example/playlistmaker/new/search/data/network/NetworkClient.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.new.search.data.network
+
+import com.example.playlistmaker.new.search.data.dto.Response
+
+interface NetworkClient {
+    fun doRequest(dto: Any): Response
+}
