+++ app/src/main/java/com/example/playlistmaker/new/search/domain/api/TrackRepository.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.new.search.domain.api
+
+import com.example.playlistmaker.new.search.domain.models.Track
+
+interface TrackRepository {
+
+    fun getTrack(): Track
+}
