+++ app/src/main/java/com/example/playlistmaker/new/search/domain/api/TracksHistoryRepository.kt
@@ -1,6 +1,6 @@
-package com.example.playlistmaker.domain.api
+package com.example.playlistmaker.new.search.domain.api
 
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.new.search.domain.models.Track
 
 interface TracksHistoryRepository {
     fun clearTrackList()
