+++ app/src/main/java/com/example/playlistmaker/new/search/domain/use_case/TracksInteractor.kt
@@ -1,0 +1,13 @@
+package com.example.playlistmaker.new.search.domain.use_case
+
+import com.example.playlistmaker.new.search.domain.api.TracksData
+import com.example.playlistmaker.new.search.domain.models.Track
+
+interface TracksInteractor {
+
+    fun searchTracks(expression: String, consumer: TrackConsumer)
+
+    interface TrackConsumer {
+        fun consume(data: TracksData<List<Track>>)
+    }
+}
