+++ app/src/main/java/com/example/playlistmaker/new/search/ui/activity/TrackAdapter.kt
@@ -2,7 +2,7 @@ package com.example.playlistmaker.new.search.ui.activity
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.new.search.domain.models.Track
 
 class TrackAdapter(private val clickListener: TrackClickListener) :
     RecyclerView.Adapter<TrackViewHolder>() {
